Website to manage the WES Automailer database. It uses jQuery, jQuery Datatables plugin for searching and sorting, and TinyMCE for HTML editing. The code itself is in C# using MVC 4.0 and Razor.
The code still needs security, and until it has it is pointing at a local database instead of a production database. The connection string for moving to prod is in the .config file, but will need to be modified to refer to "localhost." This should also resolve the remote connectivity issues.