Major Update: Updated Tournament Code create and edit pages to require selection of a calendar event and team ids for both teams. Also made these dropdown boxes. Javascript in view supports this functionality. specFRIENDS is no longer supported by League of Legends, so no more "Friends Only" spectators option for Tournament Code games. Added "Team" model. Fixed the database linkage stuff to be to a local file in the "debug" version and to the "prod" server (I use the term loosely. We need a prod server.) for the "release" version.
postget.php
Created this file to process returned values. It receives the JSON from the game server (I store the source IP address, so it will be harder for people to spoof our results, but since the system is fundamentally anonymous there is no true guarantee of security here) (it is an in-bod post message, so some normal methods of getting post data will fail on this), parses it, and stores it in the database. As a side note, PHP now works on the server.

Database work:
Created TournamentResults and TournamentResults_Player tables to hold all tournament result data. As a side note, WES now has the capability to find who did the most of X (where X is damage done, magical damage done, champions kills, sight wards placed, or any of about 40 other stats from the game), so they can create custom competitions within their tournaments. Might be a thing to mention to them, so they can differentiate their tournament service from others.